added more tests. moved to proper package
