+++ src/test/java/ru/spbau/mit/java1/ddict/DictionaryImplTest.java
@@ -1,6 +1,8 @@
-package dict;
+package ru.spbau.mit.java1.ddict;
 
 import org.junit.Test;
+import ru.spbau.mit.java1.dict.Dictionary;
+import ru.spbau.mit.java1.dict.DictionaryImpl;
 
 import static org.junit.Assert.*;
 
@@ -11,7 +13,7 @@ public class DictionaryImplTest {
     }
 
     @Test
-    public void test1() {
+    public void testPutGetContainsSize() {
         Dictionary d = instance();
         final int size = 100;
 
@@ -36,6 +38,22 @@ public class DictionaryImplTest {
         for (int i = 0; i < size; i++) {
             assertTrue(d.contains(Integer.toString(i)));
         }
+    }
+
+    @Test
+    public void testRemoveClear() {
+        Dictionary d = instance();
+        final int size = 100;
+
+        for (int i = 0; i < size; i++) {
+            assertEquals(null, d.put(Integer.toString(i), Integer.toString(1000 + i)));
+        }
+        assertEquals(d.size(), size);
+
+        for (int i = size; i < 2 * size; ++i) {
+            assertEquals(null, d.remove(Integer.toString(i)));
+        }
+        assertEquals(d.size(), size);
 
         for (int i = size / 20; i < size; ++i) {
             assertNotEquals(null, d.remove(Integer.toString(i)));
@@ -43,6 +61,11 @@ public class DictionaryImplTest {
         assertEquals(size / 20, d.size());
 
         d.clear();
+        assertEquals(0, d.size());
+        for (int i = 0; i < size; i++) {
+            assertEquals(null, d.get(Integer.toString(i)));
+            assertFalse(d.contains(Integer.toString(i)));
+        }
     }
 
 }
